Add files via upload
Basic Computer Vision in Python. The development of a theoretical and algorithmic basis to achieve automatic visual understanding, and it is concerned with the theory behind deep learning that extract information from images and videos.  For this project, you’ll get instant access to a cloud desktop with Python, Jupyter , and OpenCV pre-installed.